* in IceModel::step(), moved "dtTempAge=0.0" reset *INSIDE* a "(updateAtDepth &&   do_energy)" conditional; this fixes a major unreported bug in -skip N runs
* replaced dtTempAge (member of IceModel) with pair t_years_TempAge,dt_years_TempAge;
  this is a sub-optimal solution because this pair must be kept compatible with
  pair grid.year,dt and because couplers are called with (potentially) non-sequential
  time steps

* made test_continuity.py go faster

* "make test" result same as before

* fixes bug #17971

* I believe there are further issues in r1575 enthalpyDrainageStep() revealed by
  first examples/searise-greenland/spinup.sh run; these issues are not present
  in -cold version

+++ src/verif/iCMthermo.cc
@@ -31,6 +31,7 @@ const PetscScalar IceCompModel::LforFG = 750000; // m
 const PetscScalar IceCompModel::ApforG = 200; // m
 
 
+/*! Re-implemented so that we can add compensatory Sigma in Tests F and G. */
 PetscErrorCode IceCompModel::temperatureStep(
        PetscScalar* vertSacrCount, PetscScalar* bulgeCount) {
   PetscErrorCode  ierr;
@@ -347,7 +348,7 @@ PetscErrorCode IceCompModel::computeIceBedrockTemperatureErrors(
   ierr = my_btu->get_temp(bedrock_temp); CHKERRQ(ierr);
 
   vector<double> zblevels = bedrock_temp->get_levels();
-  int Mbz = zblevels.size();
+  int Mbz = (int)zblevels.size();
   Tbex = new PetscScalar[Mbz];
 
   switch (testname) {
